// migrations/snapshots/daa3b56ffaf1f4e1658d505cfc87c4fd5c1c705d731c19964b593c652e93a3d0/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Bookmark: {
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly noteId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Folder: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly parentFolderId: CodecTypes['pg/int4@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Note: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly folderId: CodecTypes['pg/int4@1']['output'] | null;
      readonly publishedAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly ReadingProgress: {
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly noteId: CodecTypes['pg/int4@1']['output'];
      readonly sectionName: CodecTypes['pg/text@1']['output'] | null;
      readonly progress: CodecTypes['pg/int4@1']['output'];
      readonly scrollPosition: CodecTypes['pg/int4@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
      readonly role: 'USER' | 'ADMIN';
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Bookmark: {
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly noteId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Folder: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly parentFolderId: CodecTypes['pg/int4@1']['input'] | null;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Note: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly folderId: CodecTypes['pg/int4@1']['input'] | null;
      readonly publishedAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly ReadingProgress: {
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly noteId: CodecTypes['pg/int4@1']['input'];
      readonly sectionName: CodecTypes['pg/text@1']['input'] | null;
      readonly progress: CodecTypes['pg/int4@1']['input'];
      readonly scrollPosition: CodecTypes['pg/int4@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
      readonly role: 'USER' | 'ADMIN';
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly bookmark: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly noteId: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly folder: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly parentFolderId: CodecTypes['pg/int4@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly note: {
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly folderId: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly publishedAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly reading_progress: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly noteId: CodecTypes['pg/int4@1']['output'];
      readonly progress: CodecTypes['pg/int4@1']['output'];
      readonly scrollPosition: CodecTypes['pg/int4@1']['output'] | null;
      readonly sectionName: CodecTypes['pg/text@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly role: 'USER' | 'ADMIN';
      readonly user: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly bookmark: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly noteId: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly folder: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly parentFolderId: CodecTypes['pg/int4@1']['input'] | null;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly note: {
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly folderId: CodecTypes['pg/int4@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly publishedAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly reading_progress: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly noteId: CodecTypes['pg/int4@1']['input'];
      readonly progress: CodecTypes['pg/int4@1']['input'];
      readonly scrollPosition: CodecTypes['pg/int4@1']['input'] | null;
      readonly sectionName: CodecTypes['pg/text@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly role: 'USER' | 'ADMIN';
      readonly user: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;